backend/vurrukkulluk/src/main/java/com/lydia/vurrukkulluk/service/IngredientServiceImpl.java
package com.lydia.vurrukkulluk.service;

import com.lydia.vurrukkulluk.model.Ingredient;
import com.lydia.vurrukkulluk.repository.IngredientRepository;
import lombok.AllArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@AllArgsConstructor
public class IngredientServiceImpl implements IngredientService{

    @Autowired
    private IngredientRepository ingredientRepository;
    @Override
    public Ingredient saveIngredient(Ingredient ingredient) {
        return ingredientRepository.save(ingredient);
    }

    @Override
    public List<Ingredient> getAllIngredients() {
        return ingredientRepository.findAll();
    }

    @Override
    public Ingredient getIngredientById(int id) {
        return ingredientRepository.findById(id);
    }

    @Override
    public List<Ingredient> getIngredientsRecipeId(int id) {
        return ingredientRepository.findByRecipeId(id);
    }

    @Override
    public void deleteById(int id) {
        ingredientRepository.deleteById(id);
    }
}
